perf(gpu): 💡 implement multi-tier watermark and TTL eviction for Atlas cache This change replaces the aggressive single-frame eviction strategy with a more sophisticated multi-tier approach to reduce texture thrashing and support partial repaints.

+++ painter/src/text/shaper.rs
@@ -139,7 +139,7 @@ impl TextShaper {
       for g in cluster.glyphs {
         let mut glyph = Glyph::new(
           GlyphId(g.id),
-          cluster.source.start as u32,
+          cluster.source.start,
           g.advance as i32,
           0,
           g.x as i32,
